Updated the visual design and styling
In this commit, i updated the styling of our page using css to target the different sections of the webpage to arrive at a more user friendly website. I styled the different portions of the music mixer, mainly the audio deck for dragging audio files and the play area for dropping the audio files.